// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Project {
  id: number;
  slug: string;
  url?: string | null;
  category: number | Category;
  name: string;
  description: string;
  featuredImage?: (number | null) | Media;
  images?:
    | {
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  features?:
    | {
        feature: string;
        id?: string | null;
      }[]
    | null;
  duration?: string | null;
  role?: string | null;
  techStacks?:
    | {
        stack?: string | null;
        id?: string | null;
      }[]
    | null;
  results?:
    | {
        result: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  updatedAt: string;
  createdAt: string;
}